# docs/flow-fic-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlowFic API",
    "version": "2026-03-16",
    "description": "OpenAPI spec generated from route handlers in src/app/api and supabase/functions for import into Apidog."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Next.js app"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production Next.js app"
    }
  ],
...
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin Users"
        ],
        "operationId": "getAdminUsers",
        "summary": "List users for finance admin tools",
        "security": [
          {
            "SupabaseBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsersListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
...
    "/api/internal/creator-revenue/settle": {
      "get": {
        "tags": [
          "Internal"
        ],
        "operationId": "runCreatorRevenueSettleGet",
        "summary": "Run creator revenue settlement via GET",
        "security": [
          {
            "CronBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "at",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "description": "Optional settlement timestamp override (ISO-8601)."
          }
        ],
        "responses": {
          "200": {
            "description": "Creator revenue settlement result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorRevenueSettleResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      },
      "post": {
        "tags": [
          "Internal"
        ],
        "operationId": "runCreatorRevenueSettlePost",
        "summary": "Run creator revenue settlement via POST",
        "security": [
          {
            "CronBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "at",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "required": false,
            "description": "Optional settlement timestamp override (ISO-8601)."
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "at": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "additionalProperties": false
              },
              "example": {
                "at": "2026-03-28T00:00:00.000Z"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Creator revenue settlement result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorRevenueSettleResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        }
      }
    },
...
    "/api/admin/payouts/requests": {
      "get": {
        "tags": [
          "Admin Payouts"
        ],
        "summary": "List creator payout requests",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "writerId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SupabaseBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token in the Authorization header."
      },
      "CronBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "CRON_SECRET in the Authorization header for internal cron-protected finance endpoints."
      },
      "StripeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature over the raw request body."
      }
    },
    "responses": {
      "Error400": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Error401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Error403": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Error500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "allowed": {
            "type": "boolean"
          },
          "stripeStatus": {
            "type": "integer"
          },
          "currentBalance": {
            "type": "number"
          },
          "requiredBalance": {
            "type": "number"
          },
          "requestId": {
            "type": "string"
          },
          "duplicate": {
            "type": "boolean"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": true
      },
...
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "penName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "coinBalance": {
            "type": "number"
          },
          "vipStatus": {
            "type": "string"
          },
          "vipPlanCode": {
            "type": "string",
            "nullable": true
          },
          "vipCurrentPeriodEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastSignInAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "emailConfirmedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "penName",
          "avatarUrl",
          "coinBalance",
          "vipStatus",
          "vipPlanCode",
          "vipCurrentPeriodEnd",
          "createdAt",
          "updatedAt",
          "lastSignInAt",
          "emailConfirmedAt"
        ],
        "additionalProperties": false
      },
      "AdminUsersListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "nextPage": {
                "type": "integer",
                "nullable": true
              },
              "lastPage": {
                "type": "integer",
                "nullable": true
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "nextPage",
              "lastPage"
            ]
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUser"
            }
          }
        },
        "required": [
          "success",
          "pagination",
          "users"
        ],
        "additionalProperties": false
      },
...
      "CreatorRevenueSettleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "settledCount": {
            "type": "integer",
            "minimum": 0
          },
          "movedSatang": {
            "type": "integer",
            "minimum": 0
          },
          "executedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "settledCount",
          "movedSatang",
          "executedAt"
        ],
        "additionalProperties": false
      }
    }
  }
}